+++ src/docs/previews/accordion/disabled/tailwind.svelte
@@ -30,7 +30,10 @@
 	];
 </script>
 
-<div class="mx-auto w-full max-w-md rounded-md shadow-lg" {...$root}>
+<div
+	class="mx-auto w-[18rem] max-w-full rounded-md shadow-lg sm:w-[25rem]"
+	{...$root}
+>
 	{#each items as { id, title, description, disabled }, i}
 		<div
 			melt={$item(id)}