+++ src/docs/previews/combobox/main/css.svelte
@@ -62,7 +62,7 @@
 	];
 
 	const {
-		elements: { input, menu, item },
+		elements: { input, menu, item, label },
 		states: { open, inputValue, filteredItems },
 		helpers: { isSelected },
 	} = createCombobox({
@@ -81,7 +81,7 @@
 	});
 </script>
 
-<label>
+<label melt={$label}>
 	<span>Choose your favorite book:</span>
 	<div>
 		<input melt={$input} placeholder="Best book ever" value={$inputValue} />