+++ src/docs/previews/combobox/main/tailwind.svelte
@@ -74,7 +74,7 @@
 	};
 
 	const {
-		elements: { menu, input, item },
+		elements: { menu, input, item, label },
 		states: { open, inputValue, filteredItems },
 		helpers: { isSelected },
 	} = createCombobox({
@@ -85,13 +85,17 @@
 	});
 </script>
 
-<label class="cursor-pointer">
-	<span class="block pb-1 capitalize">Choose your favorite book:</span>
+<div class="flex flex-col gap-1">
+	<!-- svelte-ignore a11y-label-has-associated-control - $label contains the 'for' attribute -->
+	<label melt={$label}>
+		<span class="block capitalize">Choose your favorite book:</span>
+	</label>
+
 	<div class="relative">
 		<input
 			melt={$input}
 			class="flex h-10 items-center justify-between rounded-md bg-white
-            px-3 pr-12 text-magnum-700"
+					px-3 pr-12 text-magnum-700"
 			placeholder="Best book ever"
 			value={$inputValue}
 		/>
@@ -103,7 +107,7 @@
 			{/if}
 		</div>
 	</div>
-</label>
+</div>
 {#if $open}
 	<ul
 		class="z-10 flex max-h-[300px] flex-col overflow-hidden rounded-md"