+++ src/docs/previews/select/main/tailwind.svelte
@@ -4,8 +4,8 @@
 	import { fly } from 'svelte/transition';
 
 	const {
-		elements: { trigger, menu, option, group, groupLabel },
-		states: { label, open },
+		elements: { trigger, menu, option, group, groupLabel, label },
+		states: { valueLabel, open },
 		helpers: { isSelected },
 	} = createSelect({
 		forceVisible: true,
@@ -17,55 +17,63 @@
 	};
 </script>
 
-<button class="trigger" melt={$trigger} aria-label="Food">
-	{$label || 'Select an option'}
-	<ChevronDown />
-</button>
-{#if $open}
-	<div class="menu" melt={$menu} transition:fly={{ duration: 100, y: -5 }}>
-		{#each Object.entries(options) as [key, arr]}
-			<div melt={$group(key)}>
-				<div class="label" melt={$groupLabel(key)}>{key}</div>
-				{#each arr as item, i}
+<div class="flex flex-col gap-1">
+	<!-- svelte-ignore a11y-label-has-associated-control - $label contains the 'for' attribute -->
+	<label class="block" melt={$label}>Favorite Flavor</label>
+	<button
+		class="flex h-10 min-w-[220px] items-center justify-between rounded-md bg-white px-3
+	py-2 text-magnum-700 transition-opacity hover:opacity-90"
+		melt={$trigger}
+		aria-label="Food"
+	>
+		{$valueLabel || 'Select a flavor'}
+		<ChevronDown />
+	</button>
+	{#if $open}
+		<div
+			class="z-10 flex max-h-[360px] flex-col
+		overflow-y-auto rounded-md bg-white
+		p-1 focus:!ring-0"
+			melt={$menu}
+			transition:fly={{ duration: 100, y: -5 }}
+		>
+			{#each Object.entries(options) as [key, arr]}
+				<div melt={$group(key)}>
 					<div
-						class="option"
-						melt={$option({ value: item, label: item, disabled: i === 0 })}
+						class="py-1 pl-4 pr-4 font-semibold capitalize text-neutral-800"
+						melt={$groupLabel(key)}
 					>
-						{#if $isSelected(item)}
-							<div class="check">
-								<Check />
-							</div>
-						{/if}
-						{item}
+						{key}
 					</div>
-				{/each}
-			</div>
-		{/each}
-	</div>
-{/if}
+					{#each arr as item}
+						<div
+							class="relative cursor-pointer rounded-md py-1 pl-8 pr-4 text-neutral-800
+						focus:z-10 focus:text-magnum-700
+					data-[highlighted]:bg-magnum-50 data-[selected]:bg-magnum-100
+					data-[highlighted]:text-magnum-900 data-[selected]:text-magnum-900"
+							melt={$option({ value: item, label: item })}
+						>
+							{#if $isSelected(item)}
+								<div class="check">
+									<Check />
+								</div>
+							{/if}
+							{item}
+						</div>
+					{/each}
+				</div>
+			{/each}
+		</div>
+	{/if}
+</div>
 
 <style lang="postcss">
-	.label {
-		@apply py-1 pl-4 pr-4  font-semibold capitalize text-neutral-800;
-	}
-
-	.menu {
-		@apply z-10 flex max-h-[360px]  flex-col overflow-y-auto;
-		@apply rounded-md bg-white p-1;
-		@apply ring-0 !important;
-	}
-	.option {
-		@apply relative cursor-pointer rounded-md py-1 pl-8 pr-4 text-neutral-800;
-		@apply focus:z-10 focus:text-magnum-700;
-		@apply data-[highlighted]:bg-magnum-50 data-[highlighted]:text-magnum-900;
-		@apply data-[selected]:bg-magnum-100 data-[selected]:text-magnum-900;
-	}
-	.trigger {
-		@apply flex h-10 min-w-[220px] items-center justify-between rounded-md bg-white px-3;
-		@apply py-2 text-magnum-700 transition-opacity hover:opacity-90;
-	}
 	.check {
-		@apply absolute left-2 top-1/2 z-20 text-magnum-500;
+		position: absolute;
+		left: theme(spacing.2);
+		top: 50%;
+		z-index: theme(zIndex.20);
 		translate: 0 calc(-50% + 1px);
+		color: theme(colors.magnum.500);
 	}
 </style>